* master:
  fix suboptimal MAX_LEN cut-off limit
  Harmonize the alpha-filter predictions at boundary
  Introduce a 'fast' alpha mode
  Fix bug for Alpha in RGBA_4444 color-mode.
  cosmetics: normalize copyright headers
  cosmetics: light include cleanup
  fix msvc build breakage after 252028a
  Some readability fixes for mux library
  Add predictive filtering option for Alpha.
  Simplify mux library code
  improve log2 test
  update Android.mk file list
  reduce number of copies and mallocs in alpha plane enc/dec
  fix some more type conversion warnings w/MSVC
  fix some uint64_t -> int conversion warnings with MSC
  simplify checks for enabling SSE2 code
  silence a msvc build warning
  vwebp: msvc build tweaks

Conflicts:
	Android.mk
	src/mux/mux.c

+++ src/dec/idec.c
@@ -1,4 +1,4 @@
-// Copyright 2011 Google Inc.
+// Copyright 2011 Google Inc. All Rights Reserved.
 //
 // This code is licensed under the same terms as WebM:
 //  Software License Agreement:  http://www.webmproject.org/license/software/
@@ -98,8 +98,8 @@ static int AppendToMemBuffer(WebPIDecoder* const idec,
   if (mem->end_ + data_size > mem->buf_size_) {  // Need some free memory
     int p;
     uint8_t* new_buf = NULL;
-    const int num_chunks = (MemDataSize(mem) + data_size + CHUNK_SIZE - 1)
-        / CHUNK_SIZE;
+    const size_t num_chunks =
+        (MemDataSize(mem) + data_size + CHUNK_SIZE - 1) / CHUNK_SIZE;
     const size_t new_size = num_chunks * CHUNK_SIZE;
     const uint8_t* const base = mem->buf_ + mem->start_;
